+++ src/common/common/parser.cpp
@@ -47,7 +47,7 @@ void delete_stop_words(
 }
 
 void split_to_ngrams(
-    std::unordered_map<std::string, int> &ngram_words,
+    std::unordered_map<std::string, size_t> &ngram_words,
     std::vector<std::string> &list_strings,
     uint16_t ngram_min_length,
     uint16_t ngram_max_length) {
@@ -77,7 +77,7 @@ void split_to_ngrams(
 void parse_text(
     std::string pars_string,
     const std::unordered_set<std::string> &stop_words,
-    std::unordered_map<std::string, int> &ngram_words,
+    std::unordered_map<std::string, size_t> &ngram_words,
     uint16_t ngram_min_length,
     uint16_t ngram_max_length) {
